Combine DescDragSourceListener and PreviewDragSourceEffect
It turns out that a DragSource listener can also set a drag preview
effect, so we don't need to have two separate classes for this; this
was slightly tricky because the drag source needs to get the image
from the preview drag source effect in order to record the right
offsets for the drag.

This changeset rolls the two classes into one, the original drag
source.

Change-Id: I249736c13e49f06886318ea64dc8de40e832d216