Add working set to New Project Wizard
I have use the code from
org.eclipse.jdt.ui.wizards.NewJavaProjectWizardPageOne
and adapt it. Suppose it's acceptable.

The Android and Android Test can be added to the same
or different Working Set.
As in the Java Project wizard, it's possible add
the project to existing working set or create new working set.

Change-Id: Ib1be35a9221bbeaf448db29cfd53fb8f1bb37c28