Icons for various widgets
Add in 16x16 icons for many of the widgets. These icons are used in
the Outline, as well as in the "Text + Icons" and "Icons Only" modes
in the palette.

The icons were generated by rendering previews with transparency on,
and layoutlib5, with the theme "Theme.Light" (since that sets dark
text labels etc). It also used a special version of the rendering
config file where for example the label of the TextView is "Abc" in
bold rather than "TextView", where there is no text label on radio and
check buttons, etc. Some icons were scaled proportionally, others just
cropped left or right, etc. Those who were cropped on the right have a
small amount of alpha-fadeout on the right to work better in the
outline.

Change-Id: I419d1c98068696a5bc650060750713e45462f01c