Update widget icons
When I generated the icons yesterday I forgot to update the background
layer; it had a flat gray background and in many cases extended a
pixel or two outside the widget foreground bounds.

This changeset updates this to have a nicer white-to-gray vertical
gradient.  It also adds icons for the <merge> and <include> tags (not
used in the palette but can show up in the outline), and tweaks the
ScrollView and LinearLayout icons.

Change-Id: I1a887f7671d099918be7e63839ac76909ff554f6