Scale drag previews with the screen zoom
If you have zoomed your layout, which is going to be common when
designing for 10" screens, then the drag preview shown when dragging
from the palette is not sized correctly. It assumes a 100% zoom.

This changeset looks at the current canvas zoom and scales the preview
image such that it matches the canvas zoom, so if you're zoomed in
then the preview will be larger, and vice versa.

Change-Id: I471b4d8ae459aad79d50a2b54ecca6bd7d7cf943