Change the JavaGenerator to handle output and dependencies.
Previously, the generator only handled a list of files to compile
and a list of files that were removed.

Now the Generator provides all that needed to do incremental
builders with known state. On top of providing a default
deltavisitor it handles file dependency and list of output
files.

The aidl generator was updated to use this, and the renderscript
generator will use the same mechanism.

Also fixed the abortOnBadSetup which through CoreException which did
nothing to stop later builders (unlike my comments said) and
poluted the workspace log with worthless messages.

Change-Id: Ib954beb9674f1387c022f926388adb8faf6cdac2