Let the Traceview plug-in open the trace file created through DDMS.
New DDMS extension: traceviewLauncher.

This allows another plug-in (traceview) to provide
a way to open traceview file. If this doesn't work
it revert to the default behavior of DDMS which is to
open the external traceview program.

also reverted the extension of traceview file
from .atv to .trace because earlier versions of
Android would automatically create those files
using this extension.

Change-Id: I2605ad47e501770ae100da2ace781b1d5d8cebc5