Add app name/icon and current render locale to the LayoutLib API.
This allows the layoutlib to render system/title/action bars
as a window decor like it would look on the device.
This can be disabled with RenderParams.setForceNoDecor().
(BTW, Params was renamed RenderParams too)

Also minor update to the API by replacing an int with Density
since the enum is now accessible to the API and layoutlib.

Change-Id: Ic37770a9276d12af90c60199a84b04cb64e7c3a1