Improve handling for GestureOverlayView
The layout editor has special handling for the "root" element, which
for example suppresses hover-highlighting for the root and a click
will initiate a marquee selection rather than a drag.

Some layouts wrap the layout inside a GestoreOverlayView. This does
not work well since the "real" root is its child, which will not be
treated as a proper root.

This changeset fixes this situation by handling this scenario such
that both the gesture overlay and its child is treated as the root.

Also make fix such that the icon in the outline also works for the
gesture overlay.

Change-Id: Ia96d2bc402958b44f9f82646ff0fc4552ab0cf14