Fix selection hint painting on honeycomb
The selection hints, which for example display the constraints when
you are in a RelativeLayout, are painted below or to the right of the
canvas. The positioning of this text was based on walking up the
ViewInfo hierarchy and taking the root element's bounds as the bounds
of the canvas image. The text was then placed below or to the right
of this image.

This no longer works with honeycomb since we now paint an action bar
and a system bar, and the bounds of these are not included in the root
ViewInfo, so as a result the selection hints were painted on top of
the system bar.

This changes the algorithm to use the image bounds itself rather than
the view info bounds.

Change-Id: I0e6e04608fb55af476c58a33785d5a7d5ac511c1