Make project-wide locale and render target settings
This changeset makes the locale and render target settings
project-wide, meaning that whenever you change the locale, or the
render target, the given locale or render target will be shown in all
layouts from the same project that you switch to.

The Locale and Render Target combo boxes are moved in to the top line
of the configuration chooser, and the dock mode and daytime mode
toggles are moved to the second line, such that the project-wide
settings are on the first line and the layout-specific settings are on
the second line.

Now, whenever you switch to a new language and open a different
layout, the language of the new layout will be forced to the current
locale. The only exception to this is for locale-specific layouts. If
you create a specific layout for a given locale, then that layout will
always be shown using its own locale, regardless of the project-wide
setting.

Change-Id: Idbe7e465a74afccfc0c68005a8784eafb2e8878d