Leave scrollbars visible
Remove the code I added recently which hides the scrollbars when they
are not enabled. While I want to restore it later, I need to guard
against a possible scenario where showing the scrollbars causes the
viewport size to change which can in turn cause the scrollbars to be
hidden recursively. For R10 let's play it safe.

Change-Id: I3fa30035ecbede55fd6e0063f222efc6fead386f