Add support for permissions and intents in declaration hyperlinks
Go to Declaration hyperlinks now work for builtin intents and
permission references, such as
  <uses-permission android:name="android.permission.CHANGE_CON...
and
  <action android:name="android.intent.action.MAIN" />
and
  <category android:name="android.intent.category.HOME" />

Jumping to the declaration will open a browser tab in the editor,
showing the relevant documentation for the given intent/permission.

If SDK documentation is installed with the SDK manager, then it will
locate the locally installed documentation, and otherwise it will
point to the corresponding page under developer.android.com.

This changeset also improves handling of references to services and
activities; in addition to supporting the short form (a dot followed
by the base name) it now also handles fully qualified names and just
basenames.

Change-Id: I2cb38023bbe16111aa1feee3a8df10112aca774b