Perform auto-zoom-to-fit on device changes
When you change the device in the configuration chooser (e.g. either
the screen type or the orientation), the layout editor will now
automatically apply a special version of the "Zoom to Fit" scaling
operation: Zoom to fit but at most 100%.

In practice this means that if you have zoomed out a lot (for example
to accomodate a tablet in portrait mode) and you switch to landscape,
it will zoom back in a bit further such that the landscape orientation
fits better without leaving a lot of unused screen space.  Similarly,
if you go from a low resolution screen to a higher resolution screen,
it will zoom out if necessary to keep the picture in view.  It will
however never zoom to more than 100%, so if you go to a lower
resolution screen will not show blurry pixels, you will see the
natural size of the image.

This changeset also contains a fix for the fit-to-zoom code such that
the margin handling works properly for zoom factors greater than one.

Change-Id: I14a6061e0492c6a116eb6d4fcedd0b5c108caabe