XML editing: add ="" for attributes and change icons
This changeset makes a couple of changes for editing XML files:

First, it changes the code completion proposals for attributes such
that when the attribute is inserted, it is inserted along with ="",
and the caret is placed inside the double quotes.

Second, it changes the icons for attributes and elements to use the
plain Eclipse attribute and element icons rather than
first-letter-derived icons. This is similar to how Eclipse does icons
for other file types, like Java (where all methods just use a green
ball for example, rather than different icons depending on the method
name.)

(An exception to this are the view elements, where we have dedicated
icons for most of the view types. For these, the individual icons are
used.)

Change-Id: I3c6927ce772265c9c8cbd2260bdc74f5ca959f56