Fill FrameLayout children on drop depending on metadata
Adds a drop handler to the FrameLaoyut rule which looks up the
metadata for the dropped child, and sets the width and/or height to
fill based on that metadata.  For example, a LinearLayout dropped into
a FrameLayout (or say a ViewFlipper which inherits from it), then the
layout will fill. A button on the other hand will not fill, and an
EditText will fill horizontally but not vertically, and so on.

Change-Id: I65a9379fa7128cc9c68d9bfd1849b0f1dc0f4a86